feat: add movie controller and endpoint exception handling

+++ backend/src/main/java/com/rayan/cinemaapi/service/MovieService.java
@@ -32,6 +32,10 @@ public class MovieService {
     }
 
     public void deleteMovie(Long id) {
+        if (!movieRepository.existsById(id)) {
+            throw new EntityNotFoundException("Movie", id);
+        }
+
         movieRepository.deleteById(id);
     }
